Add CC0 license
I have issues with pure PD because of the un-explicit legal standing.
This is as close as I can find in keeping with that spirit. The original
file is labeled as Public Domain, and passed around in repos and on
mailing lists since. Most of the repos are OFL licensed, but that
doesn't really apply to tooling!